+++ src/services/fetchers/index.ts
@@ -1,41 +1,44 @@
 import CryptoComparePriceMultiFetcher from './CryptoComparePriceMultiFetcher';
 import CryptoCompareHistoHourFetcher from './CryptoCompareHistoHourFetcher';
 import CryptoCompareHistoDayFetcher from './CryptoCompareHistoDayFetcher';
-import GVolImpliedVolatilityFetcher from './GVolImpliedVolatilityFetcher';
-import PolygonIOStockPriceFetcher from './PolygonIOStockPriceFetcher';
-import PolygonIOCryptoPriceFetcher from './PolygonIOCryptoPriceFetcher';
 import CryptoComparePriceWSFetcher from './CryptoComparePriceWSFetcher';
-import IEXEnergyFetcher from './IEXEnergyFetcher';
-import CoinmarketcapPriceFetcher from './CoinmarketcapPriceFetcher';
+
+import PolygonIOCryptoPriceFetcher from './PolygonIOCryptoPriceFetcher';
+import PolygonIOStockPriceFetcher from './PolygonIOStockPriceFetcher';
+
 import CoinmarketcapHistoHourFetcher from './CoinmarketcapHistoHourFetcher';
 import CoinmarketcapHistoDayFetcher from './CoinmarketcapHistoDayFetcher';
-import CoingeckoPriceFetcher from './CoingeckoPriceFetcher';
-import BEACPIAverageFetcher from './BEACPIAverageFetcher';
-import OnChainDataFetcher from './OnChainDataFetcher';
-import KaikoPriceStreamFetcher from './KaikoPriceStreamFetcher';
-import OptionsPriceFetcher from './OptionsPriceFetcher';
+import CoinmarketcapPriceFetcher from './CoinmarketcapPriceFetcher';
+
+import GVolImpliedVolatilityFetcher from './GVolImpliedVolatilityFetcher';
 import YearnVaultTokenPriceFetcher from './YearnVaultTokenPriceFetcher';
+import CoingeckoPriceMultiFetcher from './CoingeckoPriceMultiFetcher';
+import KaikoPriceStreamFetcher from './KaikoPriceStreamFetcher';
+import BEACPIAverageFetcher from './BEACPIAverageFetcher';
 import {UniswapPriceFetcher} from './UniswapPriceFetcher';
+import OptionsPriceFetcher from './OptionsPriceFetcher';
 import RandomNumberFetcher from './RandomNumberFetcher';
+import OnChainDataFetcher from './OnChainDataFetcher';
+import IEXEnergyFetcher from './IEXEnergyFetcher';
 
 export {
-  CryptoComparePriceWSFetcher,
-  PolygonIOStockPriceFetcher,
-  PolygonIOCryptoPriceFetcher,
   CryptoComparePriceMultiFetcher,
   CryptoCompareHistoDayFetcher,
   CryptoCompareHistoHourFetcher,
-  GVolImpliedVolatilityFetcher,
-  IEXEnergyFetcher,
-  CoingeckoPriceFetcher,
-  CoinmarketcapPriceFetcher,
+  CryptoComparePriceWSFetcher,
+  PolygonIOStockPriceFetcher,
+  PolygonIOCryptoPriceFetcher,
   CoinmarketcapHistoHourFetcher,
   CoinmarketcapHistoDayFetcher,
-  BEACPIAverageFetcher,
-  OnChainDataFetcher,
-  KaikoPriceStreamFetcher,
-  OptionsPriceFetcher,
+  CoinmarketcapPriceFetcher,
+  GVolImpliedVolatilityFetcher,
   YearnVaultTokenPriceFetcher,
+  CoingeckoPriceMultiFetcher,
+  KaikoPriceStreamFetcher,
+  BEACPIAverageFetcher,
+  OptionsPriceFetcher,
   UniswapPriceFetcher,
   RandomNumberFetcher,
+  OnChainDataFetcher,
+  IEXEnergyFetcher,
 };