+++ test/integration/FeedProcessor.test.ts
@@ -91,7 +91,7 @@ describe('FeedProcessor integration tests', () => {
     feeds = await loadFeeds('test/feeds/feeds.yaml');
   });
 
-  describe('when runnning feeds that uses HTTP', () => {
+  describe('when running feeds that uses HTTP', () => {
     feedsFetcher.forEach(({name, apiKey}) => {
       describe(`when running feeds for ${name} fetcher`, () => {
         before(function () {
@@ -111,7 +111,7 @@ describe('FeedProcessor integration tests', () => {
     });
   });
 
-  describe('when runnning feeds that uses WS', () => {
+  describe('when running feeds that uses WS', () => {
     fetcherWSNames.forEach((name) => {
       describe(`when running feeds for ${name} fetcher`, () => {
         it(`returns data for leaves for feeds with ${name} fetcher`, async () => {