+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Coingecko multi price fetcher
+- Multi-feeds processors sub-services for FeedProcessor
+
+### Removed
+- Coingecko single price fetcher
+
 ### Changed
 - Enabled FeedProcessor integration test